+++ olive/passes/__init__.py
@@ -2,8 +2,8 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from olive.passes.olive_pass import AbstractPassConfig, FullPassConfig, Pass
-from olive.passes.pass_config import PassModuleConfig, PassParamDefault
+from olive.passes.olive_pass import FullPassConfig, Pass
+from olive.passes.pass_config import AbstractPassConfig, PassModuleConfig, PassParamDefault
 
 REGISTRY = Pass.registry
