+++ olive/workflows/run/config.py
@@ -19,8 +19,7 @@ from olive.engine import Engine, EngineConfig
 from olive.engine.packaging.packaging_config import PackagingConfig
 from olive.evaluator.olive_evaluator import OliveEvaluatorConfig
 from olive.model import ModelConfig
-from olive.passes import AbstractPassConfig
-from olive.passes.pass_config import PassParamDefault
+from olive.passes.pass_config import AbstractPassConfig, PassParamDefault
 from olive.resource_path import AZUREML_RESOURCE_TYPES
 from olive.systems.system_config import SystemConfig
